Rename files to PascalCase, remove spaces from package name
