+++ PidginMacOSIntegration.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:Pidgin MacOS Integration.xcodeproj">
+      location = "self:PidginMacOSIntegration.xcodeproj">
    </FileRef>
 </Workspace>
