+++ PidginMacOSIntegration.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
